Summary:  fixes #4 - add travis badge
**Problem:**
No testing status badge.

**Analysis:**
While not really required, having a testing status badge lends a bit of confidence in the codebase.  Since this repo is using travis for testing, and travis supplies a badge, we should display it.

**Testing:**
Manual inspection of generated README page.

**Documentation:**
Addition of badge IS the documentation.