Summary:  Fixes #5 - add ascii getch
**Problem:**
getch isn't even implemented for ascii.

**Analysis:**
a basic ascii getch is required.  Add tests for it, add it, and document
it.
Add a travis config file to run the tests when a pull request is made.

**Testing:**
Added test_ugetch.py with a test for all printable ascii chars.

**Documentation:**
Updated README to add an API section with the new getkey function.
Updated internal docstrings to describe the function.