Summary: Fixes #7 - tab support
**Problem:**
No support for catching tabs

**Analysis:**
Tab-completion is a common thing to do with low-level input CLI control.
We should enable it by allowing users to catch when the tab key is
pressed.

Return 'TAB' instead of a character.  This is generally a control, along
the same lines as 'enter', and arrow keys, home key, etc.  At least,
that's the use case being targeted.  Returning a whole string is
overloading the idea of a "key", but it's lighter-weight than returning
an object you have to parse for every key.

I might still implement things that way in the end (objects), but I'm
not going to until necessary.  Using a string is still unambiguous
enough, because other characters are single characters, rather than
strings, and even multi-char unicode "characters" are not going to
overlap with the strings being used ('TAB', 'UP', 'DOWN', etc).

**Testing:**
Added a test which uses the '\t' key, as well as the control-i combo.

**Documentation:**
Added description of return value for tab.